feat(admin): Adjust UI and logic for warehouse role
This change adjusts the admin dashboard to better suit the warehouse role. It ensures that warehouse users are automatically redirected to the 'orders' tab from the 'dashboard' when logged in. Additionally, it corrects the alignment of a dropdown menu within the AdminPage component to appear on the left instead of the right, improving UI consistency.