feat: Add watermark to frame preview
Adds a watermark overlay to the frame preview component. This watermark is applied as a tiled, rotated pattern and is always visible on top of the preview. The watermark image URL is now defined in the constants and can be easily updated. This change aims to protect generated preview images.